Saves Userdata, but doesn't save images yet due to the inability of not accessing uri
The app was crashing when I was trying to load the save images so I didn't commit the code, but I think because it it trying to access a uri that the access is limited within the activity. So when you exit that activity and open the album, the app crashes. So we have to save the image instead